KIJIREPO-11. Initial commit for the base interface around CLI tools.
The interface for the model repository tools will be leveraging the bin/kiji
tools. To create a separate 'namespace' for model-repo, this introduces a tool
that will read the first argument to it and delegate to a sub-tool so that
commands like 'bin/kiji model-repo install' will work as expected.

Reviewer: shashir