KIJIREPO-15. Build an artifact packager given a list of dependencies.
Contains an interface and an implementation for building final artifacts. The
initial implementation builds a war file containing all the given
dependencies in the WEB-INF/lib folder. The interface is designed so that all
implementations need to do is ensure that a single artifact contains all the
necessary dependencies within it.

Reviewer: lsheng