KIJIREPO-17. Add dependency resolution interfaces and implementations.
In support of building a fat artifact with dependencies, it's necessary to
actually be able to resolve dependencies. Currently, there is support for "raw"
and "maven" based dependency resolution. The raw resolver simply includes files
and/or directories of jars while the maven based resolver relies on the pom.xml
to pull in the necessary runtime dependencies.

Reviewer: lsheng