KIJIREPO-12. Initial commit of the deploy tool with artifact generation and upload.
Initial upload of the deploy tool for the model repository. The initial
packaging of the model lifecycle and corresponding code is a war file with
necessary dependencies packaged in the WEB-INF/lib folder. Dependencies can be
specified either manually by listing directories/jars or by giving a pom.xml
file location and having Maven resolve the dependencies. The war file is
uploaded to the final storage location through the maven deploy-file so it's
necessary to have the Maven binary installed and on your PATH to have this work.

Reviewer: shashir, lsheng