SCORE-89. Scoring server jetty template files for lifecycles.
There are two main "template" files when deploying model lifecycles into the
scoring server:
1) The "template"
2) The "instance"

The template refers to the set of configurations necessary so that all instances
can share a single web application. These can be considered as global
configurations shared among all instances who will use the war file which the
template references. The "instance" is what is actually accessed when the user
accesses a URL for scoring.

The various files to be placed upon a new model lifecycle deployment contain
some placeholders for substituting things like the location of the model repo,
the name/version of the lifecycle being deployed etc.

Reviewer: milo, shashir