KIJIREPO-35. Fix the list tool so that it passes through valid lifecycles.
The list tool was failing to display valid lifecycles found from the Kiji table
because of the regular expression match around artifact names and versions.
Changed how the artifacts are validated to do an exact match on the artifact
name/version if either the --artifact or --version arguments was specified.
Otherwise, pass through all lifecycles that are in the model repository table.

Reviewer: shashir